+++ schemas_v2/asset.py
